+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Added
 - `pybricksdev.ble.lwp3.bytecodes` module.
 - `pybricksdev.ble.lwp3.messages` module.
+- `pybricksdev lwp3 repl` command line tool.
 
 ## [1.0.0-alpha.8] - 2021-05-18
 ## Added